RayPHP/R3S/modler: remote repository listing added

+++ odyncomp.sh
@@ -1,5 +1,11 @@
 ulimit -c 0
 make
+
+if [ "$?" != "0" ]; then
+    echo "build failed"
+    exit 1
+fi
+
 rm test
 gcc $1 -g -Wall -DFORCE_LIBRAYDIUM -o test libraydium.so -Iphp/ -Iphp/main/ -Iphp/Zend -Iphp/TSRM -Iode/include/
 sync
